Fix #22 - podcast episodes in playlists are not played
This actually is the consequence of https://github.com/spotify/web-api/issues/1528: episodes come with `is_playable=false` and `episode=false`. If this happens to a track with `spotify:episode:...` as the uri, then ignore the `is_playable` flag and give it a try anyway. Might lead to issues with episodes which really aren't playable. But let's give it a try.